Enforce UTF-8 semantics while working with assets
readFile() is reading the file (with Kernel#readFileSync()) as binary
data but implicitly converting the buffer to a utf8 string by calling
Buffer#toString() on it. Presumably, this is being done to in order to
implement paranoiaCheck, which scans the file one character at a time,
looking for non-ASCII characters. Later, this utf8 string is being
checksummed, but the call to Hash#update is implictly using binary
encoding. This is in contrast to broccoli-asset-rev, which reads the
file as binary data and calls Hash#update on the buffer (with implicit
binary encoding). If the files contain utf8 characters, the checksums
don't match!

To resolve this issue, let's read the file as a utf8 string, explicitly,
and then inform Hash#update that we are checksumming utf8 data. Now the
checksums match!

I also adjusted a call to Kernel#readFileSync() elsewhere to be
consistent (use 'utf8' instead of 'UTF-8' per the Node documentation).